fix: Update import paths for userReducer in code changes. refactor: success and cancel payments are functionally
