+++ src/shared/store/authStore.ts
